+++ KiloImportService.Api/Data/Visary/Entities/ConstructionSite.cs
@@ -11,6 +11,10 @@ public class ConstructionSite
     public string? ConstructionPermissionNumber { get; set; }
     public string? ConstructionProjectNumber { get; set; }
     public string? StageNumber { get; set; }
-    public int? FinishingMaterialId { get; set; }
+    public int? RegionId { get; set; }
+    public int? TownId { get; set; }
+    public string? Address { get; set; }
     public bool Hidden { get; set; }
+    public DateTime? Version { get; set; }
+    public int? FinishingMaterialId { get; set; }
 }